requirements_advanced.txt
flask==3.1.0
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.9
flask-socketio==5.5.1
flask-cors==5.0.0
gevent==25.9.1
nltk==3.9.1
spacy==3.8.7
tensorflow==2.20.0
keras==3.11.3
scikit-learn==1.6.1
vaderSentiment==3.3.2
optuna==4.5.0
sentence-transformers==3.4.1
gensim==4.4.0
arabic-reshaper==3.0.0
python-bidi==0.6.6
camel-tools==1.4.1
pandas==2.2.3
numpy==2.2.3
matplotlib==3.10.6
seaborn==0.13.2
torch==2.6.0
duckduckgo-search==8.1.1
langchain-community==0.3.30
pywebpush==1.14.1
cryptography==43.0.1
transformers==4.45.1
requests==2.32.3
httpx==0.27.2
joblib==1.4.2
sqlalchemy==2.0.27
flask-caching==2.1.0
celery==5.4.0
redis==5.0.4
pillow==12.0.0
langdetect==1.0.9
h5py==3.11.0
plotly==5.22.0
kaleido==0.2.1
openai==1.54.3
anthropic==0.39.1
qdrant-client==1.11.2
faiss-cpu==1.7.4
chromadb==0.5.11
sentencepiece==0.2.0
accelerate==1.1.1
bitsandbytes==0.44.1
langchain==0.3.16
langchain-openai==0.2.9
tiktoken==0.8.0
diskcache==5.6.3
datasets==3.0.0
evaluate==0.4.2
scikit-optimize==0.9.0
beautifulsoup4==4.12.3
lxml==5.3.0
chardet==5.2.0
